--- src/main/java/com/leesh/quiz/config/ApplicationConfiguration.java
@@ -1,100 +1,0 @@
-package com.leesh.quiz.config;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.leesh.quiz.domain.user.User;
-import com.leesh.quiz.domain.user.UserRepository;
-import com.leesh.quiz.security.CustomUserDetails;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.AuthenticationProvider;
-import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
-import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.web.AuthenticationEntryPoint;
-import org.springframework.security.web.access.AccessDeniedHandler;
-
-import java.util.HashMap;
-
-import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
-
-@RequiredArgsConstructor
-@Configuration
-public class ApplicationConfiguration {
-
-    private final UserRepository userRepository;
-    private final ObjectMapper objectMapper;
-
-    @Bean
-    public UserDetailsService userDetailsService() {
-        return username -> {
-
-            User user = userRepository.findByEmail(username)
-                    .orElseThrow(() -> new UsernameNotFoundException("User not found with email : " + username));
-
-            return new CustomUserDetails(user);
-        };
-    }
-
-    @Bean
-    public AuthenticationProvider authenticationProvider() {
-        DaoAuthenticationProvider authProvider = new DaoAuthenticationProvider();
-        authProvider.setUserDetailsService(userDetailsService());
-        authProvider.setPasswordEncoder(passwordEncoder());
-        return authProvider;
-    }
-
-    @Bean
-    public AuthenticationManager authenticationManager(AuthenticationConfiguration authenticationConfiguration) throws Exception {
-        return authenticationConfiguration.getAuthenticationManager();
-    }
-
-    @Bean
-    public PasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
-
-    /**
-     * 토큰이 유효하지 않거나, 유저의 비밀번호가 틀리거나 등등
-     */
-    @Bean
-    public AuthenticationEntryPoint authenticationEntryPoint() {
-
-        return (request, response, exception) -> {
-
-            response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-            response.setContentType(APPLICATION_JSON_VALUE);
-            response.setCharacterEncoding("UTF-8");
-
-            var body = new HashMap<>();
-            body.put("error", exception.getMessage());
-
-            objectMapper.writeValue(response.getWriter(), body);
-        };
-
-    }
-
-    /**
-     * 권한이 없는 경우
-     */
-    @Bean
-    public AccessDeniedHandler accessDeniedHandler() {
-
-        return (request, response, exception) -> {
-            response.setStatus(HttpServletResponse.SC_FORBIDDEN);
-            response.setContentType(APPLICATION_JSON_VALUE);
-            response.setCharacterEncoding("UTF-8");
-
-            var body = new HashMap<>();
-            body.put("error", exception.getMessage());
-
-            objectMapper.writeValue(response.getWriter(), body);
-        };
-
-    }
-}
