+++ src/main/java/com/leesh/quiz/security/SecurityConfiguration.java
@@ -1,18 +1,35 @@
 package com.leesh.quiz.security;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.leesh.quiz.domain.user.User;
+import com.leesh.quiz.domain.user.UserRepository;
 import com.leesh.quiz.security.token.TokenAuthenticationFilter;
+import com.leesh.quiz.security.token.TokenService;
+import com.leesh.quiz.security.token.jwt.JwtAuthenticationFilter;
+import com.leesh.quiz.security.token.jwt.JwtTokenService;
+import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.AuthenticationProvider;
+import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
+import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.AuthenticationEntryPoint;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.access.AccessDeniedHandler;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 
+import java.util.HashMap;
+
 import static org.springframework.boot.autoconfigure.security.servlet.PathRequest.toH2Console;
+import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 import static org.springframework.security.config.http.SessionCreationPolicy.STATELESS;
 
 @RequiredArgsConstructor
@@ -21,9 +38,8 @@ import static org.springframework.security.config.http.SessionCreationPolicy.STA
 public class SecurityConfiguration {
 
     private final TokenAuthenticationFilter tokenAuthenticationFilter;
-    private final AuthenticationProvider authenticationProvider;
-    private final AuthenticationEntryPoint authenticationEntryPoint;
-    private final AccessDeniedHandler accessDeniedHandler;
+    private final ObjectMapper objectMapper;
+    private final UserRepository userRepository;
 
     @Bean
     public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
@@ -34,8 +50,8 @@ public class SecurityConfiguration {
                 .csrf()
                 .disable()
                 .exceptionHandling()
-                .authenticationEntryPoint(authenticationEntryPoint)
-                .accessDeniedHandler(accessDeniedHandler)
+                .authenticationEntryPoint(authenticationEntryPoint())
+                .accessDeniedHandler(accessDeniedHandler())
                 .and()
                 .headers()
                 .frameOptions()
@@ -52,10 +68,78 @@ public class SecurityConfiguration {
                 .sessionManagement()
                 .sessionCreationPolicy(STATELESS)
                 .and()
-                .authenticationProvider(authenticationProvider)
+                .authenticationProvider(authenticationProvider(userRepository))
                 .addFilterBefore(tokenAuthenticationFilter, UsernamePasswordAuthenticationFilter.class);
 
         return http.build();
+    }
+
+    /**
+     * 토큰이 유효하지 않거나, 유저의 비밀번호가 틀리거나 등등
+     */
+    @Bean
+    public AuthenticationEntryPoint authenticationEntryPoint() {
+
+        return (request, response, exception) -> {
+
+            response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
+            response.setContentType(APPLICATION_JSON_VALUE);
+            response.setCharacterEncoding("UTF-8");
+
+            var body = new HashMap<>();
+            body.put("error", exception.getMessage());
+
+            objectMapper.writeValue(response.getWriter(), body);
+        };
+
+    }
+
+    /**
+     * 권한이 없는 경우
+     */
+    @Bean
+    public AccessDeniedHandler accessDeniedHandler() {
+
+        return (request, response, exception) -> {
+            response.setStatus(HttpServletResponse.SC_FORBIDDEN);
+            response.setContentType(APPLICATION_JSON_VALUE);
+            response.setCharacterEncoding("UTF-8");
+
+            var body = new HashMap<>();
+            body.put("error", exception.getMessage());
+
+            objectMapper.writeValue(response.getWriter(), body);
+        };
+
+    }
+
+    @Bean
+    public UserDetailsService userDetailsService(UserRepository userRepository) {
+        return username -> {
+
+            User user = userRepository.findByEmail(username)
+                    .orElseThrow(() -> new UsernameNotFoundException("User not found with email : " + username));
+
+            return new CustomUserDetails(user);
+        };
+    }
+
+    @Bean
+    public AuthenticationProvider authenticationProvider(UserRepository userRepository) {
+        DaoAuthenticationProvider authProvider = new DaoAuthenticationProvider();
+        authProvider.setUserDetailsService(userDetailsService(userRepository));
+        authProvider.setPasswordEncoder(passwordEncoder());
+        return authProvider;
+    }
+
+    @Bean
+    public AuthenticationManager authenticationManager(AuthenticationConfiguration authenticationConfiguration) throws Exception {
+        return authenticationConfiguration.getAuthenticationManager();
+    }
+
+    @Bean
+    public PasswordEncoder passwordEncoder() {
+        return new BCryptPasswordEncoder();
     }
 
 }
